refactor: const rename, fixed sigmoid issue
- change uninformative names like DEBUGGING_OUTPUTS
- fixed issue where logistic derivative was hardcoded into
get_partial_derivs()
- fixed issue where logistic_deriv function wasn’t what I needed it to.
Issue was that I was passing y=logistic(x) when I had coded
logistic_deriv to expect just x. Updated sigmoid_deriv functions to
accept (x, y) so that the general signature will still have x available
if we need it, but y can be treated as a convenience variable so we
don’t have to calc it twice.
- Updated forward() and get_partial_derivs() to pass around
layer_inputs so that they can be passed as x to the sigmoid_deriv
functions